+++ lib/grape/api.rb
@@ -64,8 +64,6 @@ module Grape
       def const_missing(*args)
         if base_instance.const_defined?(*args)
           base_instance.const_get(*args)
-        elsif parent && parent.const_defined?(*args)
-          parent.const_get(*args)
         else
           super
         end